+++ lib/src/other_utils/other_utils.dart
@@ -1,3 +1,4 @@
 export 'convert_types.dart';
+export 'doubly_linked_list.dart';
 export 'global_functions.dart';
 export 'time_utils.dart';